client: connection: Fix lifetime error
network_transmit_future method uses rx end of a channel to receive requests
like mqtt publish from user and sends it to network. Future combinators usually
consume the caller but that isn't suitable with the current design because we
need to reuse the rx after reconnections. So we use by_ref() method.

However, this results in the compiler not being able to infer the correct
lifetime. Fix this by specifying a lifetime.

Signed-off-by: Sanchayan Maity <maitysanchayan@gmail.com>