Sanity: Various warning, compile and clippy fixes (#138)
* Fix: Remove a clone call that creates a temporary borrowed

The clone creates a temprary that doesn't live long enough. This is
failure on at least Rust 1.32.0.

* Misc: Consolidate tokio features

Tokio allows to select certain features out of the tokio crate.

* Fix: Trivial clippy lints

* Misc: Suppress clippy warnings

Not sure how to proceed with those

* Fix: Examples compilation failures

The gcloud example relied on certificates that are not commited. This
patch adds a check for existence.

The disconnection is realized by the shutdown method of Client. The
example is renamed.

* Fix: Remove unused warnings in test target

* Feature: Add basic travis configuration

The travis configuration should also contains the nativetls features
once #137 is fixed.